Sprinting particles!

+++ src/main/kotlin/entity/player/data/Position.kt
@@ -4,7 +4,7 @@ package com.aznos.entity.player.data
  * Represents a block position in the world
  */
 data class Position(
-    var x: Int,
-    var y: Int,
-    var z: Int,
+    var x: Double,
+    var y: Double,
+    var z: Double,
 )
