+++ src/main/kotlin/packets/play/in/ClientBlockPlacementPacket.kt
@@ -31,9 +31,9 @@ class ClientBlockPlacementPacket(data: ByteArray) : Packet(data) {
         hand = input.readVarInt()
 
         val blockPosition = input.readLong()
-        val x = (blockPosition shr 38).toInt()
-        val y = (blockPosition and 0xFFF).toInt()
-        val z = (blockPosition shl 26 shr 38).toInt()
+        val x = (blockPosition shr 38).toDouble()
+        val y = (blockPosition and 0xFFF).toDouble()
+        val z = (blockPosition shl 26 shr 38).toDouble()
         location = Position(x, y, z)
 
         face = input.readVarInt()
